requirements.txt
django
django-filter
djangorestframework
drf-yasg
psycopg2-binary
spotipy
